Add more tests, fix QB emulation (unsigned comaprison), improve memory latency emulation

+++ src/main/java/com/github/vlsi/pru/plc110/RegisterField.java
@@ -17,7 +17,7 @@ public enum RegisterField {
 
   RegisterField(int bitWidth) {
     this.bitWidth = bitWidth;
-    this.bitMask = (1 << bitWidth) - 1;
+    this.bitMask = (int) ((1L << bitWidth) - 1);
   }
 
   public int getBitWidth() {
